requirements-dev.txt
pytest>=3.6.0
pytest-django==3.1.1
pytest-cov>=2.6.1
pytest-xdist>=1.26.1
flake8>=3.5.0